moved HppcMaps to o.e.common.collect package
